refactors event logic so it is shared
