+++ car/movement_listener.py
@@ -1,5 +1,6 @@
 
-from car.events import EventHandler, MovingEvent, NotMovingEvent, CarStoppedEvent
+from car.events import MovingEvent, NotMovingEvent, CarStoppedEvent
+from shared.events import EventHandler
 from haversine import haversine, Unit
 import logging
 import time
